Move ROOM related objects to a new package
