+++ app/src/main/java/com/booknerd/EchoPin/data/dao/ReminderDao.kt
@@ -1,4 +1,4 @@
-package com.booknerd.EchoPin
+package com.booknerd.EchoPin.data.dao
 
 import androidx.room.Dao
 import androidx.room.Delete
@@ -6,7 +6,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import kotlinx.coroutines.flow.Flow
+import com.booknerd.EchoPin.data.entities.Reminder
 
 @Dao
 interface ReminderDao {
